moving Secret13 to Types.
